Conflicts:
	src/main/java/com/elster/jupiter/validation/impl/ValidationRulePropertiesImpl.java

+++ coko/com.elster.jupiter.validation/src/main/java/com/elster/jupiter/validation/ValidationRuleProperties.java
@@ -1,9 +1,11 @@
 package com.elster.jupiter.validation;
 
+import java.math.BigDecimal;
+
 public interface ValidationRuleProperties {
     String getName();
 
-    long getValue();
+    BigDecimal getValue();
 
     ValidationRule getRule();
 }
