+++ coko/com.elster.jupiter.validation/src/main/java/com/elster/jupiter/validation/impl/ValidationRulePropertiesImpl.java
@@ -3,17 +3,19 @@ package com.elster.jupiter.validation.impl;
 import com.elster.jupiter.validation.ValidationRule;
 import com.elster.jupiter.validation.ValidationRuleProperties;
 
+import java.math.BigDecimal;
+
 final class ValidationRulePropertiesImpl implements ValidationRuleProperties {
 
     private String name;
-    private long value;
+    private BigDecimal value;
     private long ruleId;
 
     private transient ValidationRule rule;
 
     private ValidationRulePropertiesImpl() {}     //for persistence
 
-    public ValidationRulePropertiesImpl(ValidationRule rule, String name, long value) {
+    public ValidationRulePropertiesImpl(ValidationRule rule, String name, BigDecimal value) {
         this.rule = rule;
         this.name = name;
         this.value = value;
@@ -34,7 +36,7 @@ final class ValidationRulePropertiesImpl implements ValidationRuleProperties {
     }
 
     @Override
-    public long getValue() {
+    public BigDecimal getValue() {
         return value;
     }
